pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coralogix-mcp"
version = "0.1.0"
description = "Coralogix MCP (Monitoring and Control Panel) for log analysis and service monitoring"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "HealthifyMe", email = "dev@healthifyme.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
coralogix-mcp = "coralogix_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/healthifyme/coralogix-mcp"
Repository = "https://github.com/healthifyme/coralogix-mcp.git"
Issues = "https://github.com/healthifyme/coralogix-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["coralogix_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=coralogix_mcp --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["E501"] 